+++ src/internationalization/en.json
@@ -1,0 +1,4 @@
+{
+  "Feedback": "Feedback",
+  "Search subsets": "Search subsets"
+}
