+++ src/internationalization/nn.json
@@ -1,0 +1,4 @@
+{
+  "Feedback": "Tilbakemeldingar",
+  "Search subsets": "Søke uttrekk"
+}
