+++ src/internationalization/no.json
@@ -1,0 +1,4 @@
+{
+  "Feedback": "Tilbakemeldinger",
+  "Search subsets": "Søke uttrekk"
+}
